djblog/requirements.txt
asgiref==3.5.2
bootstrap4==0.1.0
cffi==1.15.1
Django==4.1.3
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
misaka==2.1.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
drf-yasg==1.21.7
